+++ src/lib.rs
@@ -30,8 +30,12 @@
     trivial_numeric_casts,
     unused_extern_crates,
     unused_import_braces,
-    future_incompatible
+    future_incompatible,
+    rust_2018_compatibility,
+    rust_2018_idioms,
+    clippy::all
 )]
+#![allow(clippy::verbose_bit_mask, clippy::cast_lossless)]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(feature = "use-intrinsics", feature(link_llvm_intrinsics))]
 
@@ -138,7 +142,7 @@ pub mod consts {
 impl f16 {
     /// Constructs a 16-bit floating point value from the raw bits.
     #[inline]
-    pub fn from_bits(bits: u16) -> f16 {
+    pub const fn from_bits(bits: u16) -> f16 {
         f16(bits)
     }
 
@@ -168,7 +172,7 @@ impl f16 {
 
     /// Converts an `f16` into the underlying bit representation.
     #[inline]
-    pub fn to_bits(self) -> u16 {
+    pub const fn to_bits(self) -> u16 {
         self.0
     }
 
@@ -926,6 +930,11 @@ pub mod vec {
     }
 }
 
+#[allow(
+    clippy::cognitive_complexity,
+    clippy::float_cmp,
+    clippy::neg_cmp_op_on_partial_ord
+)]
 #[cfg(test)]
 mod test {
     use super::*;